Use vnet
Expose server and client API
Remove ReadFromCM as it is not portable
Added STUN binding request test
Relates to pion/webrtc#712
