+++ README.md
@@ -108,6 +108,7 @@ Check out the [CONTRIBUTING.md](CONTRIBUTING.md) to join the group of amazing pe
 * [Sean DuBois](https://github.com/Sean-Der) - *Original Author*
 * [winds2016](https://github.com/winds2016) - *Windows platform testing*
 * [songjiayang](https://github.com/songjiayang) - *SongJiaYang*
+* [Yutaka Takeda](https://github.com/enobufs) - *vnet*
 
 ### License
 MIT License - see [LICENSE.md](LICENSE.md) for full text
