+++ cmd/simple-turn/main.go
@@ -8,29 +8,28 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/pion/logging"
 	"github.com/pion/turn"
 )
 
-type myTurnServer struct {
-	usersMap map[string]string
-}
-
-func (m *myTurnServer) AuthenticateRequest(username string, srcAddr net.Addr) (password string, ok bool) {
-	if password, ok := m.usersMap[username]; ok {
-		return password, true
+func createAuthHandler(usersMap map[string]string) turn.AuthHandler {
+	return func(username string, srcAddr net.Addr) (string, bool) {
+		if password, ok := usersMap[username]; ok {
+			return password, true
+		}
+		return "", false
 	}
-	return "", false
 }
 
 func main() {
-	m := &myTurnServer{usersMap: make(map[string]string)}
+	usersMap := map[string]string{}
 
 	users := os.Getenv("USERS")
 	if users == "" {
 		log.Panic("USERS is a required environment variable")
 	}
 	for _, kv := range regexp.MustCompile(`(\w+)=(\w+)`).FindAllStringSubmatch(users, -1) {
-		m.usersMap[kv[1]] = kv[2]
+		usersMap[kv[1]] = kv[2]
 	}
 
 	realm := os.Getenv("REALM")
@@ -40,28 +39,31 @@ func main() {
 
 	udpPortStr := os.Getenv("UDP_PORT")
 	if udpPortStr == "" {
-		log.Panic("UDP_PORT is a required environment variable")
+		udpPortStr = "3478"
 	}
 	udpPort, err := strconv.Atoi(udpPortStr)
 	if err != nil {
 		log.Panic(err)
 	}
 
-	args := turn.StartArguments{
-		Server:  m,
-		Realm:   realm,
-		UDPPort: udpPort,
-	}
-
+	var channelBindTimeout time.Duration
 	channelBindTimeoutStr := os.Getenv("CHANNEL_BIND_TIMEOUT")
 	if channelBindTimeoutStr != "" {
-		channelBindTimeout, err := time.ParseDuration(channelBindTimeoutStr)
+		channelBindTimeout, err = time.ParseDuration(channelBindTimeoutStr)
 		if err != nil {
 			log.Panicf("CHANNEL_BIND_TIMEOUT=%s is an invalid time Duration", channelBindTimeoutStr)
 		}
-
-		args.ChannelBindTimeout = channelBindTimeout
 	}
 
-	turn.Start(args)
+	s := turn.NewServer(&turn.ServerConfig{
+		Realm:              realm,
+		AuthHandler:        createAuthHandler(usersMap),
+		ChannelBindTimeout: channelBindTimeout,
+		LoggerFactory:      logging.NewDefaultLoggerFactory(),
+	})
+
+	err = s.Listen("0.0.0.0", udpPort)
+	if err != nil {
+		log.Panic(err)
+	}
 }
