+++ go.mod
@@ -4,7 +4,11 @@ go 1.12
 
 require (
 	github.com/gortc/turn v0.8.0
+	github.com/pion/logging v0.2.1
 	github.com/pion/stun v0.3.1
+	github.com/pion/transport v0.8.0
 	github.com/pkg/errors v0.8.1
+	github.com/stretchr/objx v0.2.0 // indirect
+	github.com/stretchr/testify v1.3.0
 	golang.org/x/net v0.0.0-20190403144856-b630fd6fe46b
 )
