+++ internal/allocation/channel_bind.go
@@ -1,11 +1,11 @@
 package allocation
 
 import (
-	"fmt"
 	"net"
 	"time"
 
 	"github.com/gortc/turn"
+	"github.com/pion/logging"
 )
 
 // ChannelBind represents a TURN Channel
@@ -15,18 +15,28 @@ type ChannelBind struct {
 	ID            turn.ChannelNumber
 	allocation    *Allocation
 	lifetimeTimer *time.Timer
+	log           logging.LeveledLogger
+}
+
+// NewChannelBind creates a new ChannelBind
+func NewChannelBind(id turn.ChannelNumber, peer net.Addr, log logging.LeveledLogger) *ChannelBind {
+	return &ChannelBind{
+		ID:   id,
+		Peer: peer,
+		log:  log,
+	}
 }
 
 func (c *ChannelBind) start(lifetime time.Duration) {
 	c.lifetimeTimer = time.AfterFunc(lifetime, func() {
 		if !c.allocation.RemoveChannelBind(c.ID) {
-			fmt.Printf("Failed to remove ChannelBind for %v %x %v \n", c.ID, c.Peer, c.allocation.fiveTuple)
+			c.log.Errorf("Failed to remove ChannelBind for %v %x %v", c.ID, c.Peer, c.allocation.fiveTuple)
 		}
 	})
 }
 
 func (c *ChannelBind) refresh(lifetime time.Duration) {
 	if !c.lifetimeTimer.Reset(lifetime) {
-		fmt.Printf("Failed to reset ChannelBind timer for %v %x %v \n", c.ID, c.Peer, c.allocation.fiveTuple)
+		c.log.Errorf("Failed to reset ChannelBind timer for %v %x %v", c.ID, c.Peer, c.allocation.fiveTuple)
 	}
 }
